+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 #### next release (8.9.2)
 
+- Move types/jasmine and types/jasmine-ajax to devDependencies.
 - [The next improvement]
 
 #### 8.9.1 - 2025-03-24