+++ CHANGES.md
@@ -11,6 +11,9 @@
 - Tweak `ArcGisFeatureServerCatalogItem.imageryProvider` to return undefined until metadata has finished loading
 - Fix scss theming regression by restoring webpack alias
 
+- TSify `DragDropNotification` and convert to functional component. #7543
+- [The next improvement]
+
 #### 8.9.0 - 2025-03-17
 
 - **Breaking changes:**