+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Use theme.dark for bottomBar background
 - Remove interactjs dependency #7544
 - Rewrite drag-wrapper to use terriajs dragging implementation #7544
+- TSify `SearchBoxAndResults` and convert to a functional component #7477
 - [The next improvement]
 
 #### 8.9.1 - 2025-03-24