+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Remove `viewState.lastUploadedFiles` and replace with `DragDropFile` local state. #7563
 - TSify `DragDropNotification` and convert to functional component and styled-components. #7563
 - TSify `DataCatalogTab` and convert it to functional component #7476
+- Update to shpjs 6.1.0.
 - [The next improvement]
 
 #### 8.9.1 - 2025-03-24